requirements.txt
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.5
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
